update for detail page UI updates
