+++ lib/services/api.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 
-import 'models/Item.dart';
+import 'package:swiddledick_news/models/Item.dart';
 
-const baseUrl = "https://hacker-news.firebaseio.com/v0";
+const baseUrl = 'https://hacker-news.firebaseio.com/v0';
 
 class Api {
   static Future getTopStories() {
-    var url = "$baseUrl/topstories.json";
+    var url = '$baseUrl/topstories.json';
     return http.get(url);
   }
 
@@ -17,14 +17,14 @@ class Api {
   }
 
   static Future<Item> getItem(int id) async {
-    var url = "$baseUrl/item/${id.toString()}.json";
-
+    var url = '$baseUrl/item/${id.toString()}.json';
+    print('$url');
     final response = await http.get(url);
 
     if(response.statusCode == 200) {
       return Item.fromJson(json.decode(response.body));
     } else {
-      throw Exception("Failed to load Item");
+      throw Exception('Failed to load Item');
     }
   }
 }
